Move twitter functions to util and notifications
